[Refactoring ⚙️] authUser함수 api폴더로 이동
