+++ src/components/Navbar/components/NavbarRightList/NavbarRightList.tsx
@@ -9,7 +9,7 @@ import NavbarLoggedInMenu from "./NavbarLoggedInMenu/NavbarLoggedInMenu"
 import NavbarNotLoggedInMenu from "./NavbarNotLoggedInMenu/NavbarNotLoggedInMenu"
 import MenuIcon from "@mui/icons-material/Menu"
 import useAuthUserStore from "@/stores/useAuthUserStore"
-import { LogoutAPI, POST_API } from "@/apis/Api"
+import { LogoutAPI, POST_API, authUser } from "@/apis/Api"
 import { AxiosResponse } from "axios"
 import { useNavigate } from "react-router-dom"
 import NavbarToggleMenu from "./NavbarToggleMenu/NavbarToggleMenu"
@@ -17,7 +17,7 @@ import useNavbarToggle from "@/hooks/useNavbarToggle"
 import { HandleMenuClickProps } from "../../types"
 
 const NavbarRightList = () => {
-  const { isLoggedIn, setLogin, setLogout, authUser } = useAuthUserStore()
+  const { isLoggedIn, setLogin, setLogout } = useAuthUserStore()
   const { isToggle, toggleRef, handleToggle } = useNavbarToggle()
 
   const navigate = useNavigate()
